requirements.txt
flask
python-binance
numpy
#backtrader
#TA-Lib
telethon